Prepare parser for adding transactionV1
