+++ app/src/parser_special.c
@@ -15,10 +15,11 @@
 ********************************************************************************/
 
 #include "parser_special.h"
-#include "parser_impl.h"
+#include "parser_impl_deploy.h"
 #include "parser_common.h"
 #include "parser_txdef.h"
 #include "parser.h"
+#include "parser_utils.h"
 #include "crypto.h"
 #include "zxformat.h"
 #include "app_mode.h"
