+++ app/src/runtime_arg.c
@@ -15,7 +15,7 @@
 ********************************************************************************/
 
 #include "runtime_arg.h"
-#include "parser_impl.h"
+#include "parser_impl_deploy.h"
 #include "parser_common.h"
 #include "parser_txdef.h"
 #include "parser.h"
